# specs/001-todo-app-backend/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: "Todo Application Backend API"
  description: "API for managing tasks in a multi-user todo application."
  version: "1.0.0"
paths:
  /api/v1/tasks/:
    get:
      summary: "Get All Tasks for a User"
      description: "Retrieves a list of all tasks owned by the authenticated user."
      tags:
        - tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "A list of tasks."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: "Create a New Task"
      description: "Creates a new task for the authenticated user."
      tags:
        - tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: "The created task."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'

  /api/v1/tasks/{task_id}:
    put:
      summary: "Update a Task"
      description: "Updates the title of a specific task owned by the authenticated user."
      tags:
        - tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: "The updated task."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      summary: "Delete a Task"
      description: "Deletes a specific task owned by the authenticated user."
      tags:
        - tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: "Task deleted successfully."
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/v1/tasks/{task_id}/toggle:
    put:
      summary: "Toggle Task Completion"
      description: "Toggles the completion status of a specific task."
      tags:
        - tasks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "The task with the updated status."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Buy milk"
        is_completed:
          type: boolean
          example: false
        owner_id:
          type: integer
          example: 1
    TaskCreate:
      type: object
      properties:
        title:
          type: string
          example: "Walk the dog"
      required:
        - title
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          example: "Walk the dog in the park"
      required:
        - title
  responses:
    UnauthorizedError:
      description: "Authentication is required and has failed or has not yet been provided."
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Not authenticated"
    NotFoundError:
      description: "The requested resource was not found."
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Resource not found"
    ValidationError:
      description: "The request body is invalid."
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT